Bulk of intro tutorial.
