+++ .travis.yml
@@ -74,13 +74,23 @@ install:
   - pip install codeclimate-test-reporter
 
 script:
+
+  # Runs intro tutorial notebooks
+  cd tutorials
+  - travis_wait 45 runipy intro/Intro_Tutorial_1.ipynb
+  - travis_wait 45 runipy intro/Intro_Tutorial_2.ipynb
+  - travis_wait 45 runipy intro/Intro_Tutorial_3.ipynb
+  - travis_wait 45 runipy intro/Intro_Tutorial_4.ipynb
+  - travis_wait 45 runipy intro/Intro_Tutorial_5.ipynb
+
   # Downloads data
-  - cd tutorials/disease_tagger
+  - cd ..
+  - cd disease_tagger
   - ./download_data.sh
   - cd ..
 
-  # Run notebooks end-to-end
-  - travis_wait 30 runipy disease_tagger/CDR_Tutorial_1.ipynb
+  # Runs disease tagger tutorial notebooks
+  - travis_wait 45 runipy disease_tagger/CDR_Tutorial_1.ipynb
   - travis_wait 60 runipy disease_tagger/CDR_Tutorial_2.ipynb
   - travis_wait 45 runipy disease_tagger/CDR_Tutorial_3.ipynb
   - travis_wait 45 runipy disease_tagger/CDR_Tutorial_4.ipynb
